Fixed building steps
All processes need access to all the steps, because the column matrix
generation is made by columns.
We could try generating them by lines, to reduce the memory needed to
store the steps, but writing to the dataset by columns (one word in each
line) is very slow (see
https://github.com/ejgr-mtsiw/laid-c-hdf5-mpi/tree/read_one_column_vs_calculate_one_column).
To alleviate the problem only the root in each node builds the steps
which are then shared with every process in the node.